% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryHits.R
\name{libraryHits}
\alias{libraryHits}
\title{Annotate peaks based on an in-house reference library}
\usage{
libraryHits(x, l, RT.range = 0.25, mz.range = 0.009, unique = TRUE)
}
\arguments{
\item{x}{Data frame. Experimental data. Format must be: Peak identifiers in column 1, followed by sample abundance columns, "m.z", and "RT".}

\item{l}{Data frame. Reference library. Format must be: Compound names/identifiers in column 1, "m.z", and "RT".}

\item{RT.range}{Numeric. Maximum allowable absolute retention time difference (in minutes) for a match. Default is \code{0.25}.}

\item{mz.range}{Numeric. Maximum allowable absolute mass-to-charge difference (in Da) for a match. Default is \code{0.009}.}

\item{unique}{Logical. If \code{TRUE}, when multiple experimental peaks match a single library compound, only the one closest in Retention Time is kept. Default is \code{TRUE}.}
}
\value{
A named list containing 3 structured elements:
\itemize{
  \item \code{lib_small}: Data frame mapping successful hits, displaying time deviations, compound identities, library standards, and sample abundances.
  \item \code{lib_all}: Data frame mirroring the entire input matrix structure \code{x}, with an added \code{Compound} identity column.
  \item \code{unique_names}: Character vector listing all uniquely annotated compound assignments.
}
}
\description{
Matches experimental mass-to-charge ratios (m/z) and retention times (RT) against a reference
standard library using specified tolerance windows.
}
\details{
Elements that do not match any compound within the library parameters are cleanly categorized as \code{"unknown"}.
}
