% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg.R
\docType{data}
\name{neg}
\alias{neg}
\title{LC-MS data in negative mode from MS-DIAL}
\format{
## `neg`
A data frame with 5560 rows and 51 columns:
\describe{
  \item{peak_id}{ID from MS-DIAL for particular peak}
  \item{B1-B24}{sample names from beech tree}
  \item{S1-S24}{sample names from spruce tree}
  \item{m.z}{mass to charge ratio}
  \item{RT}{retention time}

}
}
\source{
Czech Globe, Global Change Research Institute CAS
}
\usage{
neg
}
\description{
LC-MS data in negative mode from MS-DIAL
}
\keyword{datasets}
