% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFilterCV.R
\name{peakFilterCV}
\alias{peakFilterCV}
\title{Filter by CV coefficient}
\usage{
peakFilterCV(x, cut, y)
}
\arguments{
\item{x}{data frame}

\item{cut}{numeric value of the cutoff}

\item{y}{treatment info data table (optional)}
}
\value{
A reduced data frame.
}
\description{
This function provides filter by CV coefficient and can be used with or without treatment information.
}
\details{
The format of the data frame should be: Peak names|abundances of samples|m.z|RT|...

Treatment info must contain samples name, that corresponds with x in the 1. column and column treatment
(code for combination of treatments, use [treatGroup()])
}
\examples{
peakFilterCV(neg, 0.8)
peakFilterCV(neg, 0.8, t_info_group)

}
