% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFilterNA.R
\name{peakFilterNA}
\alias{peakFilterNA}
\title{Filter by missing values}
\usage{
peakFilterNA(x, cut, y)
}
\arguments{
\item{x}{data frame}

\item{cut}{numeric value of the cutoff}

\item{y}{treatment info data table (optional)}
}
\value{
A reduced data frame.
}
\description{
This function provides filter by missing values and can be used with or without treatment information.
}
\details{
The format of the data frame should be: Peak names|abundances of samples|m.z|RT|...

Treatment info must contain samples name, that corresponds with x in the 1. column and column treatment
(code for combination of treatments, use [treatGroup()])
}
\examples{
peakFilterNA(neg, 0.5)
peakFilterNA(neg, 0.5, t_info_group)

}
