% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFilterRT.R
\name{peakFilterRT}
\alias{peakFilterRT}
\title{Filter by similar RT and their correlation}
\usage{
peakFilterNA(x, dec, RTrange, cut)
}
\arguments{
\item{x}{data frame}

\item{dec}{numeric value, on how many decimal places the RT should be rounded}

\item{RTrange}{numeric value, which defines the range (+-) of similar RT}

\item{cut}{numeric value of the correlation cutoff}
}
\value{
A reduced data frame.
}
\description{
This function provides filter by RT and their correlation.
}
\details{
* The format of the data frame should be: Peak names|abundances of samples|m.z|RT|...
}
\examples{
peakFilterRT(neg,3,0.2,0.95)


}
