% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_lib.R
\docType{data}
\name{pkg_lib}
\alias{pkg_lib}
\title{In house library for this package data}
\format{
## `pkg_lib`
A data frame with 32 rows and 5 columns:
\describe{
  \item{Name}{compound name}
  \item{m.z}{mass to charge ratio}
  \item{RT}{retention time}
  \item{library}{where this information comes from}
  \item{InChIKey}{InChIKey code for the compound}

}
}
\source{
Czech Globe, Global Change Research Institute CAS
}
\usage{
pkg_lib
}
\description{
In house library for this package data
}
\keyword{datasets}
