% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\docType{data}
\name{pos}
\alias{pos}
\title{LC-MS data in positive mode from MS-DIAL}
\format{
## `pos`
A data frame with 1315 rows and 51 columns:
\describe{
  \item{peak_id}{ID from MS-DIAL for particular peak}
  \item{B1-B24}{sample names from beech tree}
  \item{S1-S24}{sample names from spruce tree}
  \item{m.z}{mass to charge ratio}
  \item{RT}{retention time}

}
}
\source{
Czech Globe, Global Change Research Institute CAS
}
\usage{
pos
}
\description{
LC-MS data in positive mode from MS-DIAL
}
\keyword{datasets}
