% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_info.R
\docType{data}
\name{t_info}
\alias{t_info}
\title{Treatment info for the package data with combined treatment code}
\format{
## `t_info`
A data frame with 48 rows and 6 columns:
\describe{
  \item{Sample}{sample name (colnames in case of neg/pos data)}
  \item{Species}{name of the specie}
  \item{CO2}{treatment info in CO2 condition (ambient, elevated)}
  \item{Water}{treatment info in water condition (watered, drought)}
  \item{Nitrogen}{treatment info in nitrogen condition (enhanced with nitrogen, control)}
  \item{Treatment}{code for all conditions in which the sample is in}

}
}
\source{
Czech Globe, Global Change Research Institute CAS
}
\usage{
t_info
}
\description{
Treatment info for the package data with combined treatment code
}
\keyword{datasets}
