% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatCases.R
\name{treatCases}
\alias{treatCases}
\title{Treatments of specific peak in complete cases.}
\usage{
treatCases(x, y, mass, RT, n)
}
\arguments{
\item{x}{data frame}

\item{y}{data frame - treatment info}

\item{mass}{peak mass (m/z)}

\item{RT}{retention time}

\item{n}{number of samples}
}
\value{
A string of the treatments code.
}
\description{
Based on m/z and RT this function will return treatments in which the complete cases are.
}
\details{
The format of the data frame x should be: Peak names|abundances of samples|m.z|RT|...

The data frame y needs column named treatment which contains the code for all treatments together (for this function [treatGroup()]
can be used). Also the first column should correspond to the names of samples in data frame x.
}
\examples{
treatCases(neg,t_info,447.12851,5.252,48)
treatCases(neg,t_info,447.12872,12.386,48)
treatCases(neg,t_info,447.22214,13.226,48)

}
