% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatGroup.R
\name{treatGroup}
\alias{treatGroup}
\title{Unique code for combined treatment}
\usage{
treatGroup(x, number.of.treatments, my_abbreviations)
}
\arguments{
\item{x}{data frame with treatment info}

\item{number.of.treatments}{integer of treatment columns (types of treatment)}

\item{my_abbreviations}{character vector of the abbreviations you want to use in alphabetical order (optional)}
}
\value{
data frame containing all original columns of \code{x} with new column with the unique treatment code
}
\description{
Based on the given abbreviation for the names of treatment, it makes an unique code for each treatment combination to which the sample is
exposed.
}
\details{
The format of the data frame should be: sample name|treatment groups|...

The vector of abbreviations has to be in alphabetical order (small cases first). If missing, then the treatment code will be formed
from the names of the treatments combined by underscore.
}
\examples{
abb <- c('B','S','A','E','D','W','n','N')
treatGroup(t_info, 4, abb)

treatGroup(t_info, 4)

}
