% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPeaks.R
\name{barPeaks}
\alias{barPeaks}
\title{Barplots of 3 lowest variable peaks based on mass (m/z).}
\usage{
barPeaks(x, mass, n)
}
\arguments{
\item{x}{data frame}

\item{mass}{peak mass (m/z)}

\item{n}{number of samples}
}
\value{
Bar plots of max 3 peaks with similar m/z and their RT and CV coefficient.
}
\description{
The function calculates CV coefficient (SD/Mean) of each peak with the given mass (m/z)
and plots abundances of all samples for the max 3 lowest CV coefficient.
}
\details{
The format of the data frame should be: Peak names|abundances of samples|m.z|RT|...

The mass should be put as a string.
}
\examples{
barPeaks(neg, '355.10', 48)

}
