% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCorr.R
\name{checkCorr}
\alias{checkCorr}
\title{Correlation of two peaks in the same retention time.}
\usage{
checkCorr(x, y = NULL, masses, RT, n)
}
\arguments{
\item{x}{data frame}

\item{y}{NULL (default) or a data frame with compatible dimensions to x. The default is equivalent to y = x}

\item{masses}{vector of peak masses}

\item{RT}{retention time}

\item{n}{number of samples}
}
\value{
Correlation of the two peaks.
}
\description{
This function gives spearman correlation coefficient of two peaks, defined by their mass (m/z).
They can be from the same data frame, or two different data frames. The combining factor is their retention time (RT).
}
\details{
The format of the data frame (x,y) should be: Peak names|abundances of samples|m.z|RT|...

The masses can be either written as numeric (rounded) or as character (not rounded).

The correlation is computed by spearman coefficiant and if there are missing values than the correlation between each pair of variables is computed using all complete pairs of observations on those variables.
}
\examples{
checkCorr(pos, masses = c(595.2, 449.08), RT = 11.8, n= 48)
checkCorr(pos, neg, masses = c(595.2, 593.15), RT = 11.8, n= 48)

}
