% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFormat.R
\name{checkFormat}
\alias{checkFormat}
\title{Data format control}
\usage{
checkFormat(x, n)
}
\arguments{
\item{x}{data frame}

\item{n}{number of samples}
}
\value{
Logical scalar with the value TRUE if everything is in order, or an
error specifying where is problem.
}
\description{
Given the data table and number of samples, the functions checks if the table
is in good format for the rest of MetaboPeak functions
}
\details{
This function validates the structure of LC-MS peak tables.

It checks whether:
\itemize{
  \item the input is a data.frame
  \item required columns ("m.z", "RT") are present
  \item the number of sample columns matches \code{n}
  \item sample abundance columns are numeric
}
}
\examples{
\donttest{
data("neg", package = "MetaboPeak")
checkFormat(neg, 48)
}
}
