% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highestAbund.R
\name{highestAbund}
\alias{highestAbund}
\title{Name of the sample in which m/z is highest.}
\usage{
highestAbund(x, mass, n)
}
\arguments{
\item{x}{data frame}

\item{mass}{peak mass (m/z)}

\item{n}{number of samples}
}
\value{
A string of the sample name.
}
\description{
Given the mass (m/z) it looks for the highest value of abundance throughout the samples and returns name of the sample.
}
\details{
The format of the data frame should be: Peak names|abundances of samples|m.z|RT|...

The mass should be put as a string.
}
\examples{
highestAbund(neg,'593.12805',48)
highestAbund(pos,'192.05482',48)

}
