% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRegion.R
\name{interestRegion}
\alias{interestRegion}
\title{Filter peaks in retention range}
\usage{
interestRegion(x, start, end)
}
\arguments{
\item{x}{data frame}

\item{start}{numeric value from which RT should the peaks be picked}

\item{end}{numeric value to which RT should the peaks be picked}
}
\value{
A reduced data frame.
}
\description{
This function reduces the original data frame to retention times that are biological relevant (without the noise at the beginning and end).
This region needs to be defined by the user based on length of their elution time and chromatograms.
}
\details{
The format of the data frame should be: Peak names|abundances of samples|m.z|RT|...
}
\examples{
neg_small <- interestRegion(neg, 1.4, 25)


}
