% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryHits.R
\name{libraryHits}
\alias{libraryHits}
\title{Annotation peaks based on in house library}
\usage{
libraryHits(x, l, RT.range, mz.range, unique)
}
\arguments{
\item{x}{data frame}

\item{l}{library (data frame)}

\item{RT.range}{numeric vector with 2 values}

\item{mz.range}{numeric vector with 2 values}

\item{unique}{logical, if there should not be replicates with the same compoud name}
}
\value{
data frame in the fromat: RT difference|Compound name|m.z|RT|abundances of samples
}
\description{
This function combines data frame with m/z and RT with given library.
Based on the ranges of m/z and RT it tries to match m/z,RT of the library with the detected.
}
\details{
* The format of the data frame (x) should be: Peak names|abundances of samples|m.z|RT|...

* The format of the library: Compound names|m.z|RT|...
}
\examples{
data_new <- libraryHits(neg, pkg_lib, RT.range = c(-0.25, 0.25), mz.range = c(-0.009, 0.009), unique = T)

}
