% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryHits.R
\name{libraryHits}
\alias{libraryHits}
\title{Annotation peaks based on in house library}
\usage{
libraryHits(x, l, RT.range, mz.range, unique)
}
\arguments{
\item{x}{data frame}

\item{l}{library (data frame)}

\item{RT.range}{numeric value determining the range of RT}

\item{mz.range}{numeric value determining the range of mz}

\item{unique}{logical, if there should not be replicates with the same compound name}
}
\value{
list of 3 objects (lib_small, lib_all and unique_name)

* lib_small = data frame in the format: RT difference|Compound name|m.z|RT|abundances of just annotated samples

* lib_all = data frame in the format: Compound name|x (all the samples, unknown = 'unknown')

* unique_names = string vector of all the compound names, that have been annotated
}
\description{
This function combines data frame with m/z and RT with given library.
Based on the ranges of m/z and RT it tries to match m/z,RT of the library with the detected.
}
\details{
The format of the data frame (x) should be: Peak names|abundances of samples|m.z|RT|...

The format of the library: Compound names|m.z|RT|...
}
\examples{
data_new <- libraryHits(neg, pkg_lib, RT.range = 0.25, mz.range = 0.009, unique = T)

}
