% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakFilter.R
\name{peakFilter}
\alias{peakFilter}
\title{Filter by NA, CV coefficient and/or correlated RT}
\usage{
peakFilter(x, y, fNA, fCV, fRT, parallel, n_cors)
}
\arguments{
\item{x}{data frame}

\item{y}{treatment info data table}

\item{fNA}{character vector}

\item{fCV}{character vector}

\item{fRT}{character vector}
}
\value{
A reduced data frame.
}
\description{
This function provides 3 types of filters. It needs the abundances table, treatment info with combined treatment code under the column
treatment (function treatGroup can be used to make this code) and some thresholds.

1) based on NA values: it calculates which peaks (rows) have more NAs than
threshold and if there are no complete cases for any treatment group, than it removes those rows.

2) based on CV coefficient: it calculates which peaks (rows) have bigger CV coefficient than
threshold and if there are no complete cases for any treatment group, than it removes those rows.

3) based on correlated RT: it rounds the RT and orders them, than based on thresholds groups similar retention times and calculates correlation
(spearman). From this correlated group it picks the one row with max m/z. (it might take some time)
}
\details{
The format of the data frame should be: Peak names|abundances of samples|m.z|RT|...

Treatment info must contain samples name, that corresponds with x in the 1. column and column treatment
(code for combination of treatments, use [treatGroup()])

fNA and fCV can be either defined as F/FALSE or with 2 argument vector ('T', threshold). fRT can be either defined as F/FALSE, or
with 4 argument vector ('T', how many decimal places should RT be rounded, range value in which to look for similar RT, correlation
threshold - how well should the rows be correlated to take them as the same)

parallel can be either defined as F/FALSE or T/TRUE

n_cores can be set as NULL and the computer will use number of available cores -1, or it can be set as any number that would work best for
used computer.
}
\examples{
peakFilter(neg, t_info_group, fNA = c('T',0.5), fCV = F, fRT = F)
peakFilter(neg, t_info_group, fNA = c('T',0.5), fCV = c('T', 0.4), fRT = F)
peakFilter(neg, t_info_group, fNA = c('T',0.5), fCV = c('T', 0.4), fRT = c('T',3,0.2,0.95)


}
