% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakInfo.R
\name{peakInfo}
\alias{peakInfo}
\title{Peak information and statistics}
\usage{
peakInfo(x, n, ID)
}
\arguments{
\item{x}{data frame}

\item{n}{number of samples}

\item{ID}{numeric (optional)}
}
\value{
A data frame with all the peak information for all or for one
}
\description{
This function returns information about individual peak on all the data table (RT, m/z, % of NAs, Max intensity, Mean, SD and CV coefficient).
If ID is provided, than it returns the same information, but for just this one peak.
}
\details{
The format of the data frame x should be: Peak names|abundances of samples|m.z|RT|...
}
\examples{
z <- peakInfo(neg,48)
z[1:5,]

peakInfo(neg,48,20)

}
