% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatGroup.R
\name{treatGroup}
\alias{treatGroup}
\title{Unique code for combined treatment}
\usage{
treatGroup(x, treatment_names, my_abbreviations, number.of.treatments)
}
\arguments{
\item{x}{data frame with treatment info}

\item{treatment_names}{character vector of the treatment names from the x}

\item{my_abbreviations}{character vector of the abbreviations you want to use}

\item{number.of.treatments}{integer of treatment columns (types of treatment)}
}
\value{
x with new column with the treatment code
}
\description{
Based on the given abbreviation for the names of treatment, it makes an unique code for each treatment combination to which the sample is
exposed.
}
\details{
* The format of the data frame should be: sample name|treatment groups|...
}
\examples{
tr <- c('S','B','AC','EC','D','W','N-','N+')
abb <- c('S','B','A','E','D','W','n','N')
treatGroup(t_info, tr, abb, 4)

}
